Demo the issue with many retry attempts
Scroll through the list. Each list item will fail to load its Source,
because the Source URI doesn't exist. It will retry many times, with a
long delay. The list items' LoadingPlaceholders point to local files that
match the index of the list item. So the first item's LoadingPlaceholder
is an image of the text "1", second item "2", etc. However, if you scroll
through the list, then you will see that the CachedImages don't update
their LoadingPlaceholder when the list items are recycled.